chore: centralize `MergeLock`-related code
This moves MergeLock-related code into a new module,
`crate::postgres::storage::merge`.

It seems sensible to keep these bits together under the storage module
as they're directly related.

This is in anticipation of some future changes to locking during merging
