+++ pg_search/src/postgres/storage/mod.rs
@@ -95,6 +95,7 @@ mod blocklist;
 pub mod buffer;
 pub mod linked_bytes;
 pub mod linked_items;
+pub mod merge;
 pub mod utils;
 
 pub use self::linked_bytes::LinkedBytesList;
